requirements.txt
django>=1.6,<1.6.99
django-ckeditor-updated
django-admin-export>=1.8
django-autocomplete-light>=2.0.0a1
django-custom-field
django-simple-import
django-dajax>=0.9.2
django-dajaxice>=0.5.5
django-grappelli>=2.4.3
django-mass-edit
django-pagination>=1.0.7
django-reversion>=1.6.3
django-localflavor
django-floppy-gumby
django-widget-tweaks
requests>=0.13.3
simplejson>=2.3.2
httpagentparser>=1.0.2
poster>=0.8.1
suds>=0.4
south>=1.0
xlwt>=0.7.2
raven>=3.3.7
reportlab
django-floppyforms>=1.0
django-report-builder
django-report-scaffold
django-responsive-dashboard
django-apptemplates>=0.0.1
django-filter>=0.5.4
# django-positions 0.5.1 may have a bug; messes up OMR question order
django-positions==0.5.0
django-bower
django-su
django-cached-field>=1.2.6
feedparser>=5.1.2
django-ldap-groups
django-rosetta>=0.7.1
-e git://github.com/zsiciarz/django-rosetta-grappelli.git#egg=django-rosetta-grappelli
django-celery>=3.1.9
django-daterange-filter
django-toolbelt
django-storages
boto
Markdown==2.4
djangorestframework==2.3.13
djangorestframework-bulk
django_compressor
django-libsass
Collectfast
#mysql-python
coffee-compressor-compiler
